Allow queuing of commands.
This means that;
a) No needed to return a command. Just call the method, that'll add it to the queue. You can keep processing after that.
b) If you call more than one, then you get them queued, and executed one after the other (in different ticks).

That will make a little more sense after we add cooldown to the commands.